# Tweak new AA implementation to get similar performance as the old one on Win32