--- a/src/rt/aaA.d
+++ b/src/rt/aaA.d
@@ -55,10 +55,7 @@ private:
         buckets = allocBuckets(sz);
         firstUsed = cast(uint) buckets.length;
         entryTI = fakeEntryTI(ti.key, ti.value);
-        immutable pad = talign(keysz, ti.value.talign) - keysz;
-        if (pad >= ushort.max)
-            assert(0, "value alignment too big");
-        keypad = cast(ushort) pad;
+        valoff = talign(keysz, ti.value.talign);
 
         import rt.lifetime : hasPostblit, unqualify;
 
@@ -75,7 +72,7 @@ private:
     uint firstUsed;
     immutable uint keysz;
     immutable uint valsz;
-    ushort keypad;
+    uint valoff;
     Flags flags;
 
     enum Flags : ubyte
@@ -199,7 +196,7 @@ private void* allocEntry(in Impl* aa, in void* pkey)
     import rt.lifetime : _d_newitemU;
     import rt.util.mem;
 
-    immutable akeysz = aa.keysz + aa.keypad;
+    immutable akeysz = aa.valoff;
     void* res = void;
     if (aa.entryTI)
         res = _d_newitemU(aa.entryTI);
@@ -359,7 +356,7 @@ extern (C) void* _aaGetY(AA* aa, const TypeInfo_AssociativeArray ti, in size_t v
 
     // found a value => return it
     if (auto p = aa.findSlotLookup(hash, pkey, ti.key))
-        return p.entry + aa.keysz + aa.keypad;
+        return p.entry + aa.valoff;
 
     auto p = aa.findSlotInsert(hash);
     if (p.deleted)
@@ -384,7 +381,7 @@ extern (C) void* _aaGetY(AA* aa, const TypeInfo_AssociativeArray ti, in size_t v
         __doPostblit(p.entry, aa.keysz, unqualify(ti.key));
     }
     // return pointer to value
-    return p.entry + aa.keysz + aa.keypad;
+    return p.entry + aa.valoff;
 }
 
 /// Get RValue for key, returns null if not present
@@ -402,7 +399,7 @@ extern (C) inout(void)* _aaInX(inout AA aa, in TypeInfo keyti, in void* pkey)
 
     immutable hash = calcHash(pkey, keyti);
     if (auto p = aa.findSlotLookup(hash, pkey, keyti))
-        return p.entry + aa.keysz + aa.keypad;
+        return p.entry + aa.valoff;
     return null;
 }
 
@@ -448,7 +445,7 @@ extern (C) inout(void[]) _aaValues(inout AA aa, in size_t keysz, in size_t valsz
     auto res = _d_newarrayU(tiValueArray, aa.length).ptr;
     auto pval = res;
 
-    immutable off = aa.keysz + aa.keypad;
+    immutable off = aa.valoff;
     foreach (b; aa.buckets[aa.firstUsed .. $])
     {
         if (!b.filled)
@@ -492,7 +489,7 @@ extern (C) int _aaApply(AA aa, in size_t keysz, dg_t dg)
     if (aa.empty)
         return 0;
 
-    immutable off = aa.keysz + aa.keypad;
+    immutable off = aa.valoff;
     foreach (b; aa.buckets)
     {
         if (!b.filled)
@@ -509,7 +506,7 @@ extern (C) int _aaApply2(AA aa, in size_t keysz, dg2_t dg)
     if (aa.empty)
         return 0;
 
-    immutable off = aa.keysz + aa.keypad;
+    immutable off = aa.valoff;
     foreach (b; aa.buckets)
     {
         if (!b.filled)
@@ -537,7 +534,7 @@ extern (C) Impl* _d_assocarrayliteralTX(const TypeInfo_AssociativeArray ti, void
 
     void* pkey = keys.ptr;
     void* pval = vals.ptr;
-    immutable off = aa.keysz + aa.keypad;
+    immutable off = aa.valoff;
     foreach (_; 0 .. length)
     {
         immutable hash = calcHash(pkey, ti.key);
@@ -584,7 +581,7 @@ extern (C) int _aaEqual(in TypeInfo tiRaw, in AA aa1, in AA aa2)
     auto uti = unqualify(tiRaw);
     auto ti = *cast(TypeInfo_AssociativeArray*)&uti;
     // compare the entries
-    immutable off = aa1.keysz + aa1.keypad;
+    immutable off = aa1.valoff;
     foreach (b1; aa1.buckets)
     {
         if (!b1.filled)
@@ -606,7 +603,7 @@ extern (C) hash_t _aaGetHash(in AA* aa, in TypeInfo tiRaw) nothrow
 
     auto uti = unqualify(tiRaw);
     auto ti = *cast(TypeInfo_AssociativeArray*)&uti;
-    immutable off = aa.keysz + aa.keypad;
+    immutable off = aa.valoff;
     auto valHash = &ti.value.getHash;
 
     size_t h;
@@ -658,7 +655,7 @@ extern (C) pure nothrow @nogc
 
     void* _aaRangeFrontValue(Range r)
     {
-        return r.buckets[r.idx].entry + r.keysz + r.keypad;
+        return r.buckets[r.idx].entry + r.valoff;
     }
 
     void _aaRangePopFront(ref Range r)