This PR tries to restore performance of the AA after changing it to "state-of-the-art" dense hash map.

Results of AA related benchmarks on my mobile i7:
```
Win32:
                   oldAA     newAA     newAA_tweaked
R bulk             0.293 s   0.328 s   0.299 s
R int              0.106 s   0.117 s   0.115 s
R obj              0.129 s   0.128 s   0.127 s
R ptr              0.110 s   0.112 s   0.106 s
R resize           0.393 s   0.420 s   0.401 s
R stomper          1.026 s   1.007 s   1.013 s
R string           0.149 s   0.142 s   0.142 s
R testgc3          1.825 s   2.201 s   1.919 s
```

Win64 benefits from these optimizations aswell:
```
Win64:
                   oldAA     newAA     newAA_tweaked
R bulk             0.444 s   0.388 s   0.365 s
R int              0.139 s   0.116 s   0.113 s
R obj              0.164 s   0.144 s   0.142 s
R ptr              0.124 s   0.117 s   0.112 s
R resize           0.457 s   0.419 s   0.395 s
R stomper          0.410 s   0.389 s   0.386 s
R string           0.115 s   0.103 s   0.099 s
R testgc3          2.632 s   2.512 s   2.420 s
```

testgc3 gets slightly faster than the old implementation without the new `mix` function.