@MartinNowak 2015-05-15
Here is a pretty simple benchmark that becomes 30% slower with this patch, my CPU has 6MB L3 cache.

```d
void main(string[] args)
{
    int[int] aa;
    foreach (_; 0 .. 10)
        foreach (i; 0 .. 1_000_000) // 16MB for bucket array
            ++aa[i];
}
```