@rainers 2015-05-15
> Here is a pretty simple benchmark that is 30% slower with this patch, my CPU has 6MB of L3 cache.

Thanks. My CPU is similar (i7-2670QM), but shows only a 10% slowdown for Win64. On Win32 I had to double the size to get simliar results. Should we add this test to the benchmark suite?

The more common use case is to have smaller arrays, though. I concede that in a program that does different stuff in between AA operations, a double cache miss is more likely to happen.