@MartinNowak 2015-05-15
> Should we add this test to the benchmark suite?

Tests that run a few seconds are somewhat unproductive. Actually the stomper test was supposed to show this, but it's magnitude is much smaller as >50% of the time is spend polluting the cache.

> The more common use case is to have smaller arrays, though. I concede that in a program that does different stuff in between AA operations, a double cache miss is more likely to happen.

It's more common, but both are common enough to matter. Can we split that off into a second pull request?