@rainers 2015-05-16
> Thanks, that's a lot of useful stuff. It's a pity that you redid the effort to split the arrays, but that's a nogo given the cost of a cache miss.

I've rebased without the split arrays. The improvements are pretty small without it, though.