@rainers 2015-05-16
> We could ditch the 32-bit hash value and only save 3 extra bits (above the mask) in the lower pointer bits.
> That'll work with the 16-byte alignment of the GC and should still detect most collisions before comparing the key, but it means recomputing the hash on growth.

Interesting idea (there are 4 unused lower bits in the entry pointer, and 16 upper bits on 64-bit).