@rainers 2015-05-16
> Tests that run a few seconds are somewhat unproductive.

If I reduce the iteration count to 4, but for 2_000_000 entries, it runs for 1.5s on my (slower) system. That's less than other benchmarks.

Apropos large number of entries: the new AA limits the number of entries to 2^^32 even on 64-bit. As the Impl struct is already slightly larger than 32 bytes, I think there is no real gain from that limitation and we should remove it.