@MartinNowak 2015-05-16
> Apropos large number of entries: the new AA limits the number of entries to 2^^32 even on 64-bit. As the Impl struct is already slightly larger than 32 bytes, I think there is no real gain from that limitation and we should remove it.

Makes sense, as long as the hot stuff fits into a single cacheline. I packed everything when there was still an inline init array.