@MartinNowak 2015-05-16
I ran dfmt on the change and replaced CRLF with LF in rt.util.mem.