@MartinNowak 2015-05-16
> The memcpy/… changes are probably a slight regression

The length is a runtime value, so that isn't entirely true.