@ibuclaw 2015-05-16
> The length is a runtime value, so that isn't entirely true.

I concur with David.  This will result in a slight regression for GDC (no cross-module inlining just yet).