@ibuclaw 2015-05-16
Also, memset is recognised as a builtin in GDC, and so already would inline as outlined in this hack at no additional cost!

@MartinNowak - Revert?