@rainers 2015-05-17
> Also, mempy/memset is recognised as a builtin in GDC, and so already would inline as outlined in this hack at no additional cost!

I tried both LDC and GDC for Win64: I don't get any inlining for memset/memcpy. What's the supposed code? If it is `rep stosb` or similar it has horrible setup time for small memory blocks. See also https://issues.dlang.org/show_bug.cgi?id=14458

Maybe we should version the code for DMD and other compilers?