@ibuclaw 2015-05-17
If you are concerned, maybe you could inline it?  newitemT used to do this:  https://github.com/D-Programming-Language/druntime/commit/79bc91b41334c1805f557ef2f1606de31c6764d1#diff-50840eb79e3292d148af41006e3af980L1115

I see no problem with libc as it usually provides SSE2/3/4 optimized functions, so should have the same overhead as the D arrayops library.