@rainers 2015-05-17
> I see no problem with libc as it usually provides SSE2/3/4 optimized functions, so should have the same overhead as the D arrayops library.

Most of the optimizations in memset/memcpy are geared towards large memory chunks, while keys are likely to be just one or two words (maybe also values). It's too bad GDC and LDC don't do cross module inlining, so they don't gain little by the "inlinable" copy and set operations. Maybe we should make them private helpers inside aaa.d, and fall back to memset/memcpy for sizes > 16.