@rainers 2015-05-17
> I ran dfmt on the change and replaced CRLF with LF in rt.util.mem.

Sorry for that. I'll try to integrate checkwhitespace into my workflow... 

Is dfmt the defacto standard style now? I don't see many modules using a space after `cast()`.