@yebblies 2015-05-17
> Sorry for that. I'll try to integrate checkwhitespace into my workflow...

It's now being run automatically on the autotester.

> Is dfmt the defacto standard style now?

No.