@rainers 2015-05-17
> It's now being run automatically on the autotester.

I don't see any target in the druntime/phobos makefiles. I think it is only checked for dmd so far, isn't it?