@MartinNowak 2015-05-17
> Is dfmt the defacto standard style now?

No, but it saves me time b/c I no longer fiddle with formatting.